+++ src/main.cpp
@@ -24,7 +24,7 @@ bool helpCalled = false;
 // Analyzing command-line arguments.
 genSettings analyzeArguments(std::vector<std::string>& arguments) {
     genSettings settings;
-    for(int i=0; i<arguments.size(); i++) {
+    for(size_t i=0; i<arguments.size(); i++) {
         if(arguments[i] == "-o" || arguments[i] == "--out") {outFile = arguments[i + 1]; i += 1;} // Output file
         else if(arguments[i] == "-np" || arguments[i] == "--noPrelude") settings.noPrelude = true; // Disables importing of std/prelude
         else if(arguments[i] == "-eml" || arguments[i] == "--emitLLVM" || arguments[i] == "-emit-llvm") settings.emitLLVM = true; // Enables output of .ll files
@@ -119,7 +119,7 @@ int main(int argc, char** argv) {
         if(options.recompileStd) {
             Compiler::initialize(outFile, outType, options, {""});
             auto stdFiles = filesInDirectory(exePath + "std");
-            for(int i=0; i<stdFiles.size(); i++) {
+            for(size_t i=0; i<stdFiles.size(); i++) {
                 if(stdFiles[i].find(".ll") == std::string::npos && stdFiles[i].find(".rave") != std::string::npos) {
                     std::string compiledFile = std::regex_replace(exePath + "std/" + stdFiles[i], std::regex("\\.rave"), ".o");
                     Compiler::compile(exePath + "std/" + stdFiles[i]);
